Moved and renamed specs to be in correct package and clarify the difference between the 2 specs
